+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.7.0] - 2019-XX-XX
-### Changed
-- Improved logging performance.
+### Added
+- Added support for XEP-0163 (Personal Eventing Protocol)
 
 ## [0.5.1] - 2019-05-03
 ### Changed